renamed tests so in correct package. removed unused imports
