+++ src/test/java/com/github/onsdigital/perkin/decrypt/MockedResponse.java
@@ -1,13 +1,12 @@
-package com.github.onsdigital.perkin.services;
+package com.github.onsdigital.perkin.decrypt;
 
 import com.github.davidcarboni.httpino.Response;
 import org.apache.http.StatusLine;
-import org.apache.poi.ss.formula.functions.T;
 
 /**
  * Created by ian on 04/04/2016.
  */
-public class MockedResponse<T> extends Response{
+public class MockedResponse<T> extends Response {
     public MockedResponse(StatusLine sl, T body) {
         super(null, null);
         this.statusLine = sl;
